[C10d] Cleanup collective sequence number.
Sequence numbers must be associated with a Work object
if we want to use it as a way to report collective progress.

The API surface change is introducing Work::getSequenceNumber, which
should eventually be exposed to python.

The bulk of this change is changing gloo to make the sequence number
be always in use and weave it to the dozens subclasses of Work.

ghstack-source-id: 8bcf775ce11e5c1f7c9dd750a1185e88b4fa9860
Pull Request resolved: https://github.com/pytorch/pytorch/pull/109136