[aotinductor] Relax the CUDAGuard device index check
Summary: Although AOTInductor only supports running on a single cuda device, it does work in the case where there is a mix of cpu and cuda ops. So instead of asserting if a CUDA index appears for the first time, we check if there is only one cuda device index. This solves https://github.com/pytorch/pytorch/issues/109655

ghstack-source-id: 592efd2659efbd2fa7628d8c8a170fba181fdb43
Pull Request resolved: https://github.com/pytorch/pytorch/pull/110030