[inductor] Actually enable typing for sizevars.py and joint_graph.py
The commit message of #107862 says it enabled mypy checking for
sizevars.py, but it seems that it neglected to update .lintrunner.toml.

New type errors appear to have crept in since then, so I've fixed them
accordingly.

A similar mistake happened with #109955 for joint_graph.py, though that
one is more recent and so hasn't had any new type errors to fix.

ghstack-source-id: 00130c0a8c063e2fdde2ec7842f54280ff62a00b
Pull Request resolved: https://github.com/pytorch/pytorch/pull/110109